2019/04/2019 major fix

+++ CoderGirl_MVCMovies/Data/IMovieRatingRepository.cs
@@ -39,7 +39,7 @@ namespace CoderGirl_MVCMovies.Data
         /// </summary>
         /// <param name="movieName"></param>
         /// <returns></returns>
-        double GetAverageRatingByMovieName(string movieName);
+        decimal GetAverageRatingByMovieName(string movieName);
 
         /// <summary>
         /// Returns a list of all the ids of saved movie ratings
